// framework/magento/schema.d.ts
export type Maybe<T> = T | null
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K]
}
export type MakeOptional<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]?: Maybe<T[SubKey]> }
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]: Maybe<T[SubKey]> }
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
}
// ...
export type Aggregation = {
  __typename?: "Aggregation"
  /** Attribute code of the aggregation group. */
  attribute_code: Scalars["String"]
  /** The number of options in the aggregation group. */
  count?: Maybe<Scalars["Int"]>
  /** The aggregation display name. */
  label?: Maybe<Scalars["String"]>
  /** Array of options for the aggregation. */
  options?: Maybe<Array<Maybe<AggregationOption>>>
}

export type AggregationOption = AggregationOptionInterface & {
  __typename?: "AggregationOption"
  /** The number of items that match the aggregation option. */
  count?: Maybe<Scalars["Int"]>
  /** Aggregation option display label. */
  label?: Maybe<Scalars["String"]>
  /** The internal ID that represents the value of the option. */
  value: Scalars["String"]
}
// ...
export type ConfigurableAttributeOption = {
  __typename?: "ConfigurableAttributeOption"
  /** The ID assigned to the attribute */
  code?: Maybe<Scalars["String"]>
  /** A string that describes the configurable attribute option */
  label?: Maybe<Scalars["String"]>
  /** A string that encodes option details. */
  uid: Scalars["ID"]
  /** A unique index number assigned to the configurable product option */
  value_index?: Maybe<Scalars["Int"]>
}
// ...
export type ConfigurableVariant = {
  __typename?: "ConfigurableVariant"
  attributes?: Maybe<Array<Maybe<ConfigurableAttributeOption>>>
  product?: Maybe<SimpleProduct>
}
// ...
export type FixedProductTax = {
  __typename?: "FixedProductTax"
  /** Amount of the FPT as a money object. */
  amount?: Maybe<Money>
  /** The label assigned to the FPT to be displayed on the frontend. */
  label?: Maybe<Scalars["String"]>
}
// ...
export type LayerFilter = {
  __typename?: "LayerFilter"
  /**
   * Array of filter items.
   * @deprecated Use Aggregation.options instead.
   */
  filter_items?: Maybe<Array<Maybe<LayerFilterItemInterface>>>
  /**
   * Count of filter items in filter group.
   * @deprecated Use Aggregation.count instead.
   */
  filter_items_count?: Maybe<Scalars["Int"]>
  /**
   * Layered navigation filter name.
   * @deprecated Use Aggregation.label instead.
   */
  name?: Maybe<Scalars["String"]>
  /**
   * Request variable name for filter query.
   * @deprecated Use Aggregation.attribute_code instead.
   */
  request_var?: Maybe<Scalars["String"]>
}
// ...
export type LayerFilterItemInterface = {
  /**
   * Count of items by filter.
   * @deprecated Use AggregationOption.count instead.
   */
  items_count?: Maybe<Scalars["Int"]>
  /**
   * Filter label.
   * @deprecated Use AggregationOption.label instead.
   */
  label?: Maybe<Scalars["String"]>
  /**
   * Value for filter request variable to be used in query.
   * @deprecated Use AggregationOption.value instead.
   */
  value_string?: Maybe<Scalars["String"]>
}
// ...
export type MediaGalleryInterface = {
  /** Whether the image is hidden from view. */
  disabled?: Maybe<Scalars["Boolean"]>
  /** The label of the product image or video. */
  label?: Maybe<Scalars["String"]>
  /** The media item's position after it has been sorted. */
  position?: Maybe<Scalars["Int"]>
  /** The URL of the product image or video. */
  url?: Maybe<Scalars["String"]>
}

/** A Money object defines a monetary value, including a numeric value and a currency code. */
export type Money = {
  __typename?: "Money"
  /** A three-letter currency code, such as USD or EUR */
  currency?: Maybe<CurrencyEnum>
  /** A number expressing a monetary value */
  value?: Maybe<Scalars["Float"]>
}
// ...
export type PriceRange = {
  __typename?: "PriceRange"
  /** The highest possible price for the product. */
  maximum_price?: Maybe<ProductPrice>
  /** The lowest possible price for the product. */
  minimum_price: ProductPrice
}
// ...
export type ProductDiscount = {
  __typename?: "ProductDiscount"
  /** The actual value of the discount. */
  amount_off?: Maybe<Scalars["Float"]>
  /** The discount expressed a percentage. */
  percent_off?: Maybe<Scalars["Float"]>
}
// ...
export type ProductImage = MediaGalleryInterface & {
  __typename?: "ProductImage"
  /** Whether the image is hidden from view. */
  disabled?: Maybe<Scalars["Boolean"]>
  /** The label of the product image or video. */
  label?: Maybe<Scalars["String"]>
  /** The media item's position after it has been sorted. */
  position?: Maybe<Scalars["Int"]>
  /** The URL of the product image or video. */
  url?: Maybe<Scalars["String"]>
}
// ...
export type ProductPrice = {
  __typename?: "ProductPrice"
  /** The price discount. Represents the difference between the regular and final price. */
  discount?: Maybe<ProductDiscount>
  /** The final price of the product after discounts applied. */
  final_price: Money
  /** The multiple FPTs that can be applied to a product price. */
  fixed_product_taxes?: Maybe<Array<Maybe<FixedProductTax>>>
  /** The regular price of the product. */
  regular_price: Money
}
// ...
export type Products = {
  __typename?: "Products"
  /** Layered navigation aggregations. */
  aggregations?: Maybe<Array<Maybe<Aggregation>>>
  /**
   * Layered navigation filters array.
   * @deprecated Use aggregations instead
   */
  filters?: Maybe<Array<Maybe<LayerFilter>>>
  /** An array of products that match the specified search criteria. */
  items?: Maybe<Array<Maybe<ProductInterface>>>
  /** An object that includes the page_info and currentPage values specified in the query. */
  page_info?: Maybe<SearchResultPageInfo>
  /** An object that includes the default sort field and all available sort fields. */
  sort_fields?: Maybe<SortFields>
  /**
   * The number of products that are marked as visible. By default, in complex
   * products, parent products are visible, but their child products are not.
   */
  total_count?: Maybe<Scalars["Int"]>
}

export type SingleProducts = {
  __typename?: "Products"
  /**
   * Layered navigation filters array.
   * @deprecated Use aggregations instead
   */
  filters?: Maybe<Array<Maybe<LayerFilter>>>
  /** An array of products that match the specified search criteria. */
  items?: Maybe<Array<Maybe<ProductInterface>>>
}
// ...
export type SearchResultPageInfo = {
  __typename?: "SearchResultPageInfo"
  /** Specifies which page of results to return */
  current_page?: Maybe<Scalars["Int"]>
  /** Specifies the maximum number of items to return */
  page_size?: Maybe<Scalars["Int"]>
  /** Total pages */
  total_pages?: Maybe<Scalars["Int"]>
}
// ...
export type SortField = {
  __typename?: "SortField"
  /** Label of sort field. */
  label?: Maybe<Scalars["String"]>
  /** Attribute code of sort field. */
  value?: Maybe<Scalars["String"]>
}

/** SortFields contains a default value for sort fields and all available sort fields. */
export type SortFields = {
  __typename?: "SortFields"
  /** Default value of sort fields. */
  default?: Maybe<Scalars["String"]>
  /** Available sort fields. */
  options?: Maybe<Array<Maybe<SortField>>>
}